Put util files in one folder and extract sqrt algorithm into algorithms.tcc
