+++ libff/algebra/curves/bn128/bn128_gt.hpp
@@ -11,7 +11,7 @@
 
 #include "depends/ate-pairing/include/bn.h"
 
-#include <libff/algebra/fields/prime_base/field_utils.hpp>
+#include <libff/algebra/field_utils/field_utils.hpp>
 #include <libff/algebra/fields/prime_base/fp.hpp>
 
 namespace libff {
