+++ libff/algebra/field_utils/algorithms.tcc
@@ -3,7 +3,7 @@
 
  Implementation of interfaces for (square-and-multiply) exponentiation.
 
- See exponentiation.hpp .
+ See algorithms.hpp .
 
  *****************************************************************************
  * @author     This file is part of libff, developed by SCIPR Lab
@@ -46,6 +46,61 @@ template<typename FieldT>
 FieldT power(const FieldT &base, const unsigned long exponent)
 {
     return power<FieldT>(base, bigint<1>(exponent));
+}
+
+template<typename FieldT>
+FieldT tonelli_shanks_sqrt(const FieldT &value)
+{
+    FieldT one = FieldT::one();
+
+    size_t v = FieldT::s;
+    FieldT z = FieldT::nqr_to_t;
+    FieldT w = value^FieldT::t_minus_1_over_2;
+    FieldT x = value * w;
+    FieldT b = x * w; // b = (*this)^t
+
+#if DEBUG
+    // check if square with euler's criterion
+    FieldT check = b;
+    for (size_t i = 0; i < v-1; ++i)
+    {
+        check = check.squared();
+    }
+    if (check != one)
+    {
+        assert(0);
+    }
+#endif
+
+    // compute square root with Tonelli--Shanks
+    // (does not terminate if not a square!)
+
+    while (b != one)
+    {
+        size_t m = 0;
+        FieldT b2m = b;
+        while (b2m != one)
+        {
+            /* invariant: b2m = b^(2^m) after entering this loop */
+            b2m = b2m.squared();
+            m += 1;
+        }
+
+        int j = v-m-1;
+        w = z;
+        while (j > 0)
+        {
+            w = w.squared();
+            --j;
+        } // w = z^2^(v-m-1)
+
+        z = w.squared();
+        b = b * z;
+        x = x * w;
+        v = m;
+    }
+
+    return x;
 }
 
 } // libff
