+++ libff/algebra/fields/binary/gf256.hpp
@@ -11,7 +11,7 @@
 #include <cstddef>
 #include <cstdint>
 #include <vector>
-#include <libff/algebra/fields/bigint.hpp>
+#include <libff/algebra/field_utils/bigint.hpp>
 
 namespace libff {
