+++ libff/algebra/fields/prime_extension/fp3.tcc
@@ -10,7 +10,7 @@
 #ifndef FP3_TCC_
 #define FP3_TCC_
 
-#include <libff/algebra/fields/prime_base/field_utils.hpp>
+#include <libff/algebra/field_utils/field_utils.hpp>
 
 namespace libff {
 
@@ -236,56 +236,7 @@ Fp3_model<n,modulus> Fp3_model<n,modulus>::Frobenius_map(unsigned long power) co
 template<mp_size_t n, const bigint<n>& modulus>
 Fp3_model<n,modulus> Fp3_model<n,modulus>::sqrt() const
 {
-    Fp3_model<n,modulus> one = Fp3_model<n,modulus>::one();
-
-    size_t v = Fp3_model<n,modulus>::s;
-    Fp3_model<n,modulus> z = Fp3_model<n,modulus>::nqr_to_t;
-    Fp3_model<n,modulus> w = (*this)^Fp3_model<n,modulus>::t_minus_1_over_2;
-    Fp3_model<n,modulus> x = (*this) * w;
-    Fp3_model<n,modulus> b = x * w; // b = (*this)^t
-
-#if DEBUG
-    // check if square with euler's criterion
-    Fp3_model<n,modulus> check = b;
-    for (size_t i = 0; i < v-1; ++i)
-    {
-        check = check.squared();
-    }
-    if (check != one)
-    {
-        assert(0);
-    }
-#endif
-
-    // compute square root with Tonelli--Shanks
-    // (does not terminate if not a square!)
-
-    while (b != one)
-    {
-        size_t m = 0;
-        Fp3_model<n,modulus> b2m = b;
-        while (b2m != one)
-        {
-            /* invariant: b2m = b^(2^m) after entering this loop */
-            b2m = b2m.squared();
-            m += 1;
-        }
-
-        int j = v-m-1;
-        w = z;
-        while (j > 0)
-        {
-            w = w.squared();
-            --j;
-        } // w = z^2^(v-m-1)
-
-        z = w.squared();
-        b = b * z;
-        x = x * w;
-        v = m;
-    }
-
-    return x;
+    return tonelli_shanks_sqrt(*this);
 }
 
 template<mp_size_t n, const bigint<n>& modulus>
