+++ libff/algebra/fields/prime_extension/fp6_2over3.tcc
@@ -9,7 +9,7 @@
 
 #ifndef FP6_2OVER3_TCC_
 #define FP6_2OVER3_TCC_
-#include <libff/algebra/fields/prime_base/field_utils.hpp>
+#include <libff/algebra/field_utils/field_utils.hpp>
 #include <libff/algebra/scalar_multiplication/wnaf.hpp>
 
 namespace libff {
