+++ libff/algebra/fields/prime_field.hpp
@@ -13,7 +13,7 @@
  *             and contributors (see AUTHORS).
  * @copyright  MIT license (see LICENSE file)
  *****************************************************************************/
-#include <libff/algebra/fields/bigint.hpp>
+#include <libff/algebra/field_utils/bigint.hpp>
 #include <vector>
 
 namespace libff {
