+++ libff/algebra/scalar_multiplication/multiexp.tcc
@@ -18,8 +18,8 @@
 #include <cassert>
 #include <type_traits>
 
-#include <libff/algebra/fields/bigint.hpp>
-#include <libff/algebra/fields/prime_base/fp_aux.tcc>
+#include <libff/algebra/field_utils/bigint.hpp>
+#include <libff/algebra/field_utils/fp_aux.tcc>
 #include <libff/algebra/scalar_multiplication/multiexp.hpp>
 #include <libff/algebra/scalar_multiplication/wnaf.hpp>
 #include <libff/common/profiling.hpp>
