+++ libff/common/double.cpp
@@ -14,7 +14,7 @@
 
 #include <math.h>
 
-#include <libff/algebra/fields/bigint.hpp>
+#include <libff/algebra/field_utils/bigint.hpp>
 #include <libff/common/double.hpp>
 
 namespace libff {
