+++ libff/common/double.hpp
@@ -14,7 +14,7 @@
 
 #include <complex>
 
-#include <libff/algebra/fields/bigint.hpp>
+#include <libff/algebra/field_utils/bigint.hpp>
 
 namespace libff {
